api/pyproject.toml
[tool.poetry]
name = "CostsMap"
version = "3.0.0"
description = "CostsMap is an application for costs accounting"
authors = ["artemowkin <artyom.groshev2017@gmail.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.75.2"}
gunicorn = "^20.1.0"
aiopg = "^1.3.3"
SQLAlchemy = "^1.4.35"
databases = "^0.5.5"
passlib = "^1.7.4"
python-jose = "^3.3.0"
asyncpg = "^0.25.0"
pytest = "^7.1.1"
aiosqlite = "^0.17.0"
python-dateutil = "^2.8.2"
async-lru = "^1.0.3"
orm = {extras = ["postgresql"], version = "^0.3.1"}

[tool.poetry.dev-dependencies]
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
